lots of fixes
