+++ packages/cbioportal-utils/src/index.tsx
@@ -20,6 +20,9 @@ export * from './model/MyVariantInfo';
 export * from './model/OncoKB';
 export * from './model/RemoteData';
 
+export { default as SuperAgentClient } from './client/SuperAgentWrapped';
+export * from './types/types';
+
 export * from './mutation/MutationAnnotator';
 export * from './mutation/MutationUtils';
 export * from './mutation/ProteinChangeUtils';
