+++ packages/cbioportal-utils/src/types/types.ts
@@ -1,0 +1,9 @@
+export type SiteErrorMode = 'dialog' | 'screen' | 'alert';
+
+export type SiteError = {
+    errorObj: any;
+    dismissed: boolean;
+    title?: string;
+    mode?: SiteErrorMode;
+    customMessage?: string;
+};
