+++ src/shared/api/cbioportalInternalClientInstance.ts
@@ -1,7 +1,9 @@
 import { CBioPortalAPIInternal } from 'cbioportal-ts-api-client';
-import { extendCBioPortalAPI } from 'shared/lib/extendCBioPortalAPI';
+import { addErrorHandlingtoAPIClient } from 'shared/lib/addErrorHandlingtoAPIClient';
 
-const ExtendedCBioPortalAPI = extendCBioPortalAPI(CBioPortalAPIInternal);
+const ExtendedCBioPortalAPI = addErrorHandlingtoAPIClient(
+    CBioPortalAPIInternal
+);
 
 const internalClient = new ExtendedCBioPortalAPI();
